graphs/iso: Move SolveUnique to solve.go.
