+++ iso/solve.go
@@ -1,6 +1,11 @@
 package iso
 
-import "github.com/mewkiz/pkg/errutil"
+import (
+	"fmt"
+	"sort"
+
+	"github.com/mewkiz/pkg/errutil"
+)
 
 // SetPair marks the given node pair as known by removing it from c and storing
 // it in m. As the graph node name is no longer a valid candidate it is removed
@@ -56,4 +61,42 @@ func (eq *Equation) Dup() *Equation {
 	}
 
 	return &Equation{c: c, m: m}
+}
+
+// SolveUnique tries to locate a unique node pair in c. If successful the node
+// pair is removed from c and stored in m. As the graph node name of the node
+// pair is no longer a valid candidate it is removed from all other node pairs
+// in c.
+func (eq *Equation) SolveUnique() (ok bool, err error) {
+	// Sort keys to make the algorithm deterministic.
+	var snames []string
+	for sname := range eq.c {
+		snames = append(snames, sname)
+	}
+	sort.Strings(snames)
+
+	for _, sname := range snames {
+		candidates := eq.c[sname]
+		if len(candidates) == 1 {
+			gname := firstKey(candidates)
+			err := eq.SetPair(sname, gname)
+			if err != nil {
+				return false, errutil.Err(err)
+			}
+			return true, nil
+		}
+	}
+
+	return false, nil
+}
+
+// firstKey returns the only key in m.
+func firstKey(m map[string]bool) string {
+	if len(m) != 1 {
+		panic(fmt.Sprintf("invalid map length; expected 1, got %d", len(m)))
+	}
+	for key := range m {
+		return key
+	}
+	panic("unreachable")
 }
